backend_api/requirement.txt
Flask>=2.3
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
python-dotenv
Werkzeug